backend/requirements.txt
fastapi
uvicorn
requests
gspread
oauth2client
python-dotenv